Move everything in CvImpl to RawImpl
