+++ cscore/src/main/java/edu/wpi/first/cscore/raw/RawSink.java
@@ -22,7 +22,7 @@ public class RawSink extends ImageSink {
    * @param name Source name (arbitrary unique identifier)
    */
   public RawSink(String name) {
-    super(CameraServerJNI.createRawSink(name));
+    super(CameraServerJNI.createRawSink(name, false));
   }
 
   /**
